+++ model/scikit/dffml_model_scikit/scikit_base.py
@@ -127,7 +127,8 @@ class Scikit(Model):
     def _filename(self):
         return os.path.join(
             self.config.directory,
-            hashlib.sha384(self.config.predict.NAME.encode()).hexdigest() + ".json",
+            hashlib.sha384(self.config.predict.NAME.encode()).hexdigest()
+            + ".json",
         )
 
     async def __aenter__(self) -> "Scikit":