+++ model/tensorflow/tests/test_dnnr.py
@@ -52,7 +52,7 @@ class TestDNN(AsyncTestCase):
                 steps=1000,
                 epochs=30,
                 hidden=[10, 20, 10],
-                predict=DefFeature("TARGET",float,1),
+                predict=DefFeature("TARGET", float, 1),
                 features=cls.features,
             )
         )