+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The latest release of the documentation now includes a link to the
   documentation for the master branch (on GitHub pages).
 ### Changed
+- `predict` parameter of all model config classes has been changed from `str` to `Feature`.
 - Defining features on the command line no longer requires that defined features
   be prefixed with `def:`
 - The model predict operation will now raise an exception if the model it is